chore(one-on-one): minor improvement
